+++ packages/carp_survey_package/example/lib/example.dart
@@ -26,7 +26,7 @@ void main() async {
   protocol.addTriggeredTask(
       DelayedTrigger(delay: Duration(seconds: 30)),
       RPAppTask(
-          type: SurveyUserTask.WHO5_SURVEY_TYPE,
+          type: SurveyUserTask.SURVEY_TYPE,
           name: 'WHO-5 Survey',
           rpTask: who5Task)
         ..measures.add(Measure(type: DeviceSamplingPackage.DEVICE))